Move the severity config validation/fixing to severity.lua
