+++ lua/trouble/severity.lua
@@ -1,6 +1,50 @@
 -- severity filters
 
+local DiagnosticSeverity = vim.lsp.protocol.DiagnosticSeverity
+
 local M = {}
+
+----------------------------------
+-- config validation --
+----------------------------------
+
+-- this is a bit overboard to validate these settings, but fun
+local to_severity = function(severity)
+  if not severity then return nil end
+  return type(severity) == 'string' and DiagnosticSeverity[severity] or severity
+end
+local severity_keys = vim.tbl_keys(DiagnosticSeverity)
+local severity_names = vim.tbl_filter(function(a) return type(a) == "string" end, severity_keys)
+table.sort(severity_names, function (a, b) return to_severity(a) < to_severity(b) end)
+local severity_names_joined = table.concat(severity_names, ", ")
+local severity_expected = "nil, number in range 1..=4, or {"..severity_names_joined .. "}"
+function sev_validate(s)
+  -- Diagnostics
+  return vim.tbl_contains(severity_keys, s)
+end
+function opt_sev_validate(s)
+  if s == nil then return true end
+  if type(s) == 'number' then return s end
+  return sev_validate(s)
+end
+
+function M.fix_config(opts)
+  vim.validate {
+    min_severity = { opts.min_severity, opt_sev_validate, severity_expected },
+    cascading_severity = { opts.cascading_severity, opt_sev_validate, severity_expected },
+  }
+  -- make them 1..=4 or nil
+  opts.min_severity = to_severity(opts.min_severity)
+  -- min_severity being Hint just runs a no-op filter, so ignore it
+  if opts.min_severity == 4 then
+    opts.min_severity = nil
+  end
+  opts.cascading_severity = to_severity(opts.cascading_severity)
+end
+
+----------------------------------
+-- handling keybindings/actions --
+----------------------------------
 
 -- counter-intuitive. LSP severities are backwards.
 -- nil is least severe (no filter), 4 is next smallest (Hint), 1 is most severe (Error)
@@ -49,17 +93,18 @@ function M.__run_tests()
   print("trouble.severity tests passed")
 end
 
+-- for Trouble.action
 local severity_actions = {
   "incr_min_severity",
   "decr_min_severity",
   "incr_cascading_severity",
   "decr_cascading_severity",
 }
-
 function M.handles_action(action)
   return vim.tbl_contains(severity_actions, action)
 end
 
+-- for Trouble.action
 function M.apply_action(action, config)
   local fn = nil
   local key = nil
@@ -73,4 +118,32 @@ function M.apply_action(action, config)
   config.options[key] = fn(prev_sev)
 end
 
+--------------------------
+-- filtering severities --
+--------------------------
+
+-- Note that LSP severities are backwards in the sense that "Error"=1, "Warning"=2, "Info"=3, "Hint"=4.
+-- So "Hint" is the highest number but not the most severe.
+-- When we say min_severity in the code, it means we are filtering by <= the LSP severity number.
+
+-- must pass an integer severity to these
+function eq_severity_fn(severity)
+  return function(t) return t.severity == severity end
+end
+function min_severity_fn(severity)
+  -- recall LSP DiagnosticSeverity being backwards
+  return function(t) return t.severity <= severity end
+end
+
+--- Filters using a function. Returns a new table, and the number of diagnostics left out
+---
+--- @param items Diagnostic[]
+--- @return Diagnostic[], number
+function filter_diags(items, filter_fn)
+  local filtered = vim.tbl_filter(filter_fn, items)
+  local hidden = #items - #filtered
+  return filtered, hidden
+end
+
 return M
+
